feat(desktop): agent queue, memory, skills, UI ownership, DeepSeek
Ship per-agent serial task queues and working memory.md, skills catalog
and firewall, single-home Autopilot/runtime settings, true left-nav page
sections, deploy and Grok CLI test freezes fixed, and DeepSeek as a
meeting-brain provider. Keep meeting brains as chat LLMs and execution
runtimes (including OpenClaw/CLI) separate.

+++ soulcorp-desktop/src-tauri/src/scrum/command_center.rs
@@ -29,8 +29,8 @@ pub struct CommandCenterOverview {
     pub throttled_runs: u32,
     pub unassigned_sprint_tasks: u32,
     pub active_sprint_name: Option<String>,
-    pub burndown_remaining: u8,
-    pub burndown_total: u8,
+    pub burndown_remaining: u32,
+    pub burndown_total: u32,
     pub execution_paused: bool,
     pub alerts: Vec<CommandCenterAlert>,
 }
